spec: test locations with doubles
